AIS: Changed module to use types of known type.
Think some problems observed was due to different length of int on
STM32F103 and ATMEGA328

+++ AIS.h
@@ -11,12 +11,12 @@ public:
 	static const int msg_max = 60;
 	bool getdata(unsigned int begin, unsigned int cnt, uint8_t *data, bool isigned = false);
 
-	unsigned long get_mmsi();
-	unsigned int get_SOG();
-	unsigned int get_COG();
-	unsigned int get_HDG();
-	long get_latitude();
-	long get_longitude();
+	uint32_t get_mmsi();
+	uint16_t get_SOG();
+	uint16_t get_COG();
+	uint16_t get_HDG();
+	int32_t get_latitude();
+	int32_t get_longitude();
 	int8_t get_rot();  // Rate of Turn
 	uint8_t get_repeat();
 	uint8_t get_navStatus();
@@ -29,9 +29,9 @@ public:
 private:
 	void decode(unsigned int fillBits);
 	int getbit(unsigned int idx);
-	unsigned int get_u16(unsigned start, unsigned len);
-	unsigned long get_u32(unsigned start, unsigned len);
-	long get_i32(unsigned start, unsigned len);
+	uint16_t get_u16(unsigned start, unsigned len);
+	uint32_t get_u32(unsigned start, unsigned len);
+	int32_t get_i32(unsigned start, unsigned len);
 	int8_t get_i8(unsigned start, unsigned len);
 	uint8_t get_u8(unsigned start, unsigned len);
